still working on clubs

+++ src/components/book/LibraryList.js
@@ -6,58 +6,67 @@ import {BookClubJoinContext} from "../bookClubJoin/BookClubJoinProvider"
 
 export const LibraryList = (props) => {
     const {books, getBooks, deleteFromUserLibrary, getUserBooks, userBooks} = useContext(BookContext)
-    const {JoinABookClub} = useContext(BookClubJoinContext) || {}
+    const {CreateABookClub} = useContext(BookClubJoinContext) || {}
 
     //const [userBooks, setUserBooks] = useState([])
 
     useEffect(() => {
         getBooks()
-    }, [userBooks])
+    }, [])
 
     useEffect(() => {
         getUserBooks()
     },[])
-
+    
     return (
         <article className="libraryBooks">
             {
                 userBooks.map(libraryBook => { 
-                    const matched = books.find(b => b.booktag === libraryBook.bookId) 
-                    return <section className="libraryBook" key={libraryBook.id}>
-                        <img src={matched.cover} />
-                        <h4 className="library__title">{matched.title}</h4>
-                        <div className="library__author">{matched.author}</div>
-                        <div className="library__pages">{matched.pages}</div>
-                        <button type="submit" id={libraryBook.id}
-                            onClick={e => {
-                                
-                                e.preventDefault()
-                                const bookId = matched.id
-                                const name = matched.title
-                                JoinABookClub ({
-                                    bookId,
-                                    name
-                                })                                
-                        }}
-                        className="joinBookClubBtn">
-                        Join a Book club
-                        </button>
-                        <button type="submit" id={libraryBook.id}
-                            onClick={e => {
-                                const id = parseInt(e.target.id)
-                                
-                            deleteFromUserLibrary(id)  
-                                
-                                
-                                
-                                
-                        }}
-                        className="deleteLibraryBook">
-                            Delete From Library
-                        </button>
-                    </section>
-            
+                    if(libraryBook.userId === parseInt(localStorage.getItem("bookclub_user"))) {
+                        const matched = books.find(b => b.booktag === libraryBook.bookId) ||  {}
+                        return <section className="libraryBook" key={libraryBook.id}>
+                            <h4 className="library__title">{matched.title}</h4>
+                            <img src={matched.cover} alt="No Cover"/>
+                            <div className="library__author">Author: {matched.author}</div>
+                            <div className="library__pages">Pages: {matched.pages}</div>
+                            <button type="submit" id={libraryBook.id}
+                                onClick={e => {
+                                    
+                                    e.preventDefault()
+                                    const bookId = matched.id
+                                    const name = matched.title
+                                    CreateABookClub ({
+                                        bookId,
+                                        name
+                                    }).then(props.history.push("/clubs"))                        
+                            }}
+                            className="joinBookClubBtn">
+                            Create a Book Club
+                            </button>
+                            <button type="submit" id={libraryBook.id}
+                                onClick={e => {
+                                    const id = parseInt(e.target.id)
+                                    
+                                deleteFromUserLibrary(id)  
+                                    
+                                    
+                                    
+                                    
+                            }}
+                            className="deleteLibraryBook">
+                                Delete from Library
+                            </button>
+                        </section>
+                
 
+                    } else {
+                        return (
+                            <section className="libraryBook">
+                            
+                            </section>
+
+                        )
+                    }
                 })
